requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
--index-url https://pypi.python.org/simple

beautifulsoup4==4.9.3
    # via -r requirements.in
certifi==2020.12.5
    # via requests
chardet==4.0.0
    # via requests
facenet-pytorch==2.5.2
    # via -r requirements.in
idna==2.10
    # via requests
numpy==1.20.1
    # via
    #   -r requirements.in
    #   facenet-pytorch
    #   torch
    #   torchvision
pillow==8.1.2
    # via
    #   -r requirements.in
    #   facenet-pytorch
    #   torchvision
requests==2.25.1
    # via
    #   -r requirements.in
    #   facenet-pytorch
soupsieve==2.2.1
    # via beautifulsoup4
torch==1.8.0
    # via torchvision
torchvision==0.9.0
    # via facenet-pytorch
tqdm==4.59.0
    # via -r requirements.in
typing-extensions==3.7.4.3
    # via torch
urllib3==1.26.4
    # via requests
